refactor : S3Service class body : 공통된 S3 업로드 및 이미지 리사이징 관련 S3Service class에서 관리하여 유지 보수 및 가독성을 높임
